+++ .gitignore
@@ -63,3 +63,6 @@ node_modules
 static/sw.js*
 static/workbox*
 static/fonts
+
+.next
+*.tsbuildinfo
