+++ package.json
@@ -1,49 +1,35 @@
 {
   "name": "@dhedegaard/pollen",
-  "version": "0.1.3",
-  "main": "build/index.js",
+  "version": "1.0.0",
   "license": "GPL-3.0-or-later",
   "private": false,
   "engines": {
-    "node": ">=18",
+    "node": ">=20",
     "npm": ">=9"
   },
   "scripts": {
-    "prebuild": "rimraf ./build",
-    "build": "tsc",
-    "postbuild": "workbox generateSW workbox-config.js",
-    "watch": "tsx watch ./src/index.ts",
-    "start": "tsc && node ./build/index.js",
-    "prepublishOnly": "npm run build"
+    "build": "next build",
+    "dev": "next dev",
+    "start": "next start",
+    "lint": "next lint"
   },
   "dependencies": {
-    "@emotion/react": "^11.11.0",
+    "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.0",
-    "@mui/material": "^5.13.3",
-    "compression": "^1.7.4",
-    "express": "^4.18.2",
-    "helmet": "^6.0.0",
-    "morgan": "^1.10.0",
+    "@mui/material": "^5.15.11",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "tsx": "^3.12.7",
-    "typescript": "^5.1.3",
-    "utility-types": "^3.10.0",
-    "xml2js": "^0.6.0",
-    "zod": "^3.21.4"
+    "typescript": "^5.3.3",
+    "utility-types": "^3.11.0",
+    "xml2js": "^0.6.2",
+    "zod": "^3.22.4"
   },
   "devDependencies": {
-    "@tsconfig/node-lts": "^18.12.2",
-    "@tsconfig/strictest": "^2.0.1",
-    "@types/compression": "^1.7.2",
-    "@types/express": "^4.17.17",
-    "@types/morgan": "^1.9.4",
-    "@types/node": "^18.16.16",
-    "@types/react": "^18.2.8",
-    "@types/react-dom": "^18.2.4",
-    "@types/xml2js": "^0.4.11",
-    "rimraf": "^5.0.1",
-    "tslint-config-prettier": "^1.18.0",
-    "workbox-cli": "^7.0.0"
+    "@tsconfig/node-lts": "^20.1.1",
+    "@tsconfig/strictest": "^2.0.3",
+    "@types/node": "^20.11.24",
+    "@types/react": "^18.2.61",
+    "@types/react-dom": "^18.2.19",
+    "@types/xml2js": "^0.4.14"
   }
 }
