--- src/api.ts
@@ -1,3 +1,0 @@
-import express from 'express'
-
-export default (app: express.Express) => app
