--- workbox-config.js
@@ -1,14 +1,0 @@
-module.exports = {
-  globDirectory: 'static/',
-  globPatterns: ['**/*.{png,ico,json,js,woff2}'],
-  swDest: 'static/sw.js',
-  clientsClaim: true,
-  skipWaiting: true,
-  inlineWorkboxRuntime: true,
-  runtimeCaching: [
-    {
-      urlPattern: /\/$/,
-      handler: 'NetworkFirst',
-    },
-  ],
-}
